% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancers.R
\name{add_exclusivity}
\alias{add_exclusivity}
\title{Add exclusivity of phylogroups}
\usage{
add_exclusivity(tg, similarity)
}
\arguments{
\item{tg}{A tidygenomes object}

\item{similarity}{An expression defining a genome similarity measure}
}
\value{
A tidygenomes object
}
\description{
This function calculates the minimum similarity within phylogroups and the
maximum similarity between phylogroups, assesses exclusivity and adds these
measures to the phylogroup table.
}
