% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importers.R
\name{add_genome_metadata}
\alias{add_genome_metadata}
\title{Add genome metadata to tidygenomes object}
\usage{
add_genome_metadata(tg, genome_metadata, by = "genome")
}
\arguments{
\item{tg}{A tidygenomes object}

\item{genome_metadata}{A table having at least one variable in common with
the genomes table of the tidygenomes object}

\item{by}{The variable to join by}
}
\value{
A tidygenomes object
}
\description{
This function joins a table containing genome data to the genome table of a
tidygenomes object.
}
