% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancers.R
\name{add_orthogroup_measures}
\alias{add_orthogroup_measures}
\title{Add various orthogroup measures to the orthogroup table}
\usage{
add_orthogroup_measures(tg)
}
\arguments{
\item{tg}{A tidygenomes object}
}
\value{
A tidygenomes object
}
\description{
This function adds the variables \code{og_genes} and \code{og_genomes} to the
orthogroups table of a tidygenomes object.
}
