% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancers.R
\name{add_patterns}
\alias{add_patterns}
\title{Add presence/absence patterns to a tidygenomes object}
\usage{
add_patterns(tg)
}
\arguments{
\item{tg}{A tidygenomes object}
}
\value{
A tidygenomes object
}
\description{
This function determines unique orthogroup presence/absence patterns in
genomes and adds this information to a tidygenomes object.
}
\details{
The following will be added to the tidygenomes object:
\itemize{
\item A variable \code{pattern} in the orthogroups table, with for each orthogroup its
presence/absence pattern.
\item A table \code{components}, listing the genomes defining each presence/absence
pattern.
\item A table \code{patterns} with for each pattern, the number of orthogroups that
follow it.
}
}
\examples{
genes <- data.frame(
  gene = c("gene A", "gene B", "gene C"),
  genome = c("genome 1", "genome 1", "genome 2"),
  orthogroup = c("og alpha", "og beta", "og alpha")
)
tg <- as_tidygenomes(genes)
tg <- add_patterns(tg)

}
