% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancers.R
\name{add_phylogroup_color}
\alias{add_phylogroup_color}
\title{Add color variable for phylogroups}
\usage{
add_phylogroup_color(tg, n = 12)
}
\arguments{
\item{tg}{A tidygenomes object}

\item{n}{The number of colors}
}
\value{
A tidygenomes object
}
\description{
This function adds a variable phylogroup_color to the phylogroup table that
can be used to plot phylogroup colors in cases where there are too many
phylogroups for the color scale you want to use.
}
\details{
The variable phylogroup_color assigns a number to each phylogroup, in order
of the ladderized tree. If there are more phylogroups than numbers, the
numbers will be recycled. The goal is that two phylogroups that are directly
next to each other in the tree labels will never have the same color.
}
