% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancers.R
\name{add_phylogroup_measures}
\alias{add_phylogroup_measures}
\title{Add various phylogroup measures to the phylogroup table}
\usage{
add_phylogroup_measures(tg)
}
\arguments{
\item{tg}{A tidygenomes object}
}
\value{
A tidygenomes object
}
\description{
This function adds the variable \code{pg_genomes} to the orthogroups table of a
tidygenomes object.
}
