% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importers.R
\name{add_phylogroups}
\alias{add_phylogroups}
\title{Add phylogroups to a tidygenomes object}
\usage{
add_phylogroups(tg, phylogroups, genome_identifier = genome)
}
\arguments{
\item{tg}{A tidygenomes object}

\item{phylogroups}{A table specifying phylogroups}

\item{genome_identifier}{The name of a variable from the genomes table that
is used to identify genomes in the phylogroups table}
}
\value{
A tidygenomes object
}
\description{
Phylogroups are defined as clades in a phylogenetic tree. This function will
add user-defined phylogroups to a tidygenomes object containing a tree.
}
\details{
The phylogroups table should consist of one row per phylogroup and should
contain the following variables:
\itemize{
\item phylogroup: the name of the phylogroup
\item genome_type: the type genome of the phylogroup
\item genome_peripheral: a peripheral genome of the phylogroup
}

A phylogroup is then defined as the most recent common ancestor of the type
and peripheral genomes in the tree, along with all its descendants.

The following will be added to the tidygenomes object:
\itemize{
\item A variable \code{phylogroup} in the nodes table.
\item A variable \code{is_phylogroup_ancestor} in the nodes table.
\item A variable \code{is_phylogroup_type} in the genomes table.
}
}
