% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_rootbranch}
\alias{add_rootbranch}
\title{Add a branch to the root of the tree}
\usage{
add_rootbranch(tree, branch_length = 1)
}
\arguments{
\item{tree}{A phylogeny of class \code{phylo}, with edge lengths}

\item{branch_length}{The length of the root branch}
}
\value{
An object of class \code{phylo}
}
\description{
This function adds a branch to the root of the tree, adding one extra node in
the process.
}
