% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importers.R
\name{add_tidygenomes}
\alias{add_tidygenomes}
\title{Add a tidygenomes object to a tidygenomes object}
\usage{
add_tidygenomes(tg, data)
}
\arguments{
\item{tg}{A tidygenomes object}

\item{data}{A tidygenomes object or data type that can be converted into a
tidygenomes object}
}
\value{
A tidygenomes object
}
\description{
This function adds a second tidygenomes object, or data from which a
tidygenomes object can be created, to a tidygenomes object.
}
\details{
See \code{\link[=as_tidygenomes]{as_tidygenomes()}} for a list of what \code{data} can be.
}
\examples{
genomes <- data.frame(genome = c("genome 1", "genome 2"))
genes <- data.frame(
  gene = c("gene A", "gene B", "gene C"),
  genome = c("genome 1", "genome 1", "genome 2"),
  orthogroup = c("og alpha", "og beta", "og alpha")
)
tg <- as_tidygenomes(genomes)
tg <- add_tidygenomes(genes)

}
