% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importers.R
\name{as_tidygenomes}
\alias{as_tidygenomes}
\title{Create a new tidygenomes object}
\usage{
as_tidygenomes(data)
}
\arguments{
\item{data}{A data type that can be converted into a tidygenomes object}
}
\value{
A tidygenomes object: this is a list consisting of a genome table and
optionally extra data objects related to the genomes in the genome table
}
\description{
This function creates a new tidygenomes object from a genome table, pangenome
table, genome tree or genome pair table.
}
\details{
\code{data} should be one of the following:
\itemize{
\item a \strong{genome table}: contains a variable \code{genome}, with unique values
\item a \strong{pangenome table}: contains the variables \code{gene}, \code{genome} and
\code{orthogroup}
\item a \strong{genome tree}: should be of class \code{phylo}
\item a \strong{genome pair table}: contains the variables \code{genome_1} and
\code{genome_2}, with only unique combinations and without self-pairs
}
}
\examples{
genomes <- data.frame(genome = c("genome 1", "genome 2"))
tg <- as_tidygenomes(genomes)

}
