% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computers.R
\name{calculate_heap_coefficient}
\alias{calculate_heap_coefficient}
\title{Estimates the Heap coefficient for a pangenome}
\usage{
calculate_heap_coefficient(tg, permutations = 10, plot = FALSE)
}
\arguments{
\item{tg}{A tidygenomes object}

\item{permutations}{How many times the original data will be sampled for a range of genome numbers.}

\item{plot}{Whether to return a plot showing the datapoints and curve used to calculate the coefficient.
Defaults to true.}
}
\value{
A genome table
}
\description{
This function samples a large amount of orthogroups per genome and then fits
a function through the resulting ratios to determine the Heap's coefficient
using formula P = kN^lambda
with P the amount of orthogroups, N the amount of genomes,
k an empiric parameter depending on the pangenome and lambda the Heap's coefficient.
}
