% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancers.R
\name{collapse_species}
\alias{collapse_species}
\title{Collapse species pangenomes}
\usage{
collapse_species(tg, core_threshold = 0.9)
}
\arguments{
\item{tg}{A tidygenomes object}

\item{core_threshold}{The minimum percentage of genomes within a species
where a gene needs to be present to be considered a core gene}
}
\value{
A tidygenomes object
}
\description{
This function collapses each species to a single genome, and each orthogroup
within a species to a single gene.
}
\details{
A variable "species" must be present in the genome table for this function to
work.

A variable "status" will be added to the gene table, indicating for each gene
whether the orthogroup that gene belongs to is core or accessory within the
species.
}
