% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{complete_pairs}
\alias{complete_pairs}
\title{Complement the pairs of a pair table}
\usage{
complete_pairs(pairs, object_1, object_2)
}
\arguments{
\item{pairs}{Data frame where each row represents a pair of objects}

\item{object_1}{Variable defining the first object (unquoted)}

\item{object_2}{Variable defining the second object (unquoted)}
}
\value{
A complemented table
}
\description{
For each unique pair of objects (a, b), the function adds a row for the pair
(b, a) with the same information.
}
