% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{filter_genomes}
\alias{filter_genomes}
\title{Filter the genomes of a tidygenomes object}
\usage{
filter_genomes(tg, ...)
}
\arguments{
\item{tg}{A tidygenomes object}

\item{...}{
  Arguments passed on to \code{\link[dplyr:filter]{dplyr::filter}}
  \describe{
    \item{\code{.data}}{A data frame, data frame extension (e.g. a tibble), or a
lazy data frame (e.g. from dbplyr or dtplyr). See \emph{Methods}, below, for
more details.}
    \item{\code{.by}}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Optionally, a selection of columns to
group by for just this operation, functioning as an alternative to \code{\link[dplyr:group_by]{group_by()}}. For
details and examples, see \link[dplyr:dplyr_by]{?dplyr_by}.}
    \item{\code{.preserve}}{Relevant when the \code{.data} input is grouped.
If \code{.preserve = FALSE} (the default), the grouping structure
is recalculated based on the resulting data, otherwise the grouping is kept as is.}
  }}
}
\value{
A tidygenomes object
}
\description{
This function applies \code{\link[dplyr:filter]{dplyr::filter()}} to the genomes table of a tidygenomes
object and subsequently filters all other components of the tidygenomes
object to retain only the requested genomes. It is also possible to filter on
variables of the phylogroups table, if present.
}
\examples{
genomes <- data.frame(genome = c("genome 1", "genome 2"))
tg <- as_tidygenomes(genomes)
tg <- tg \%>\% filter_genomes(genome == "genome 1")

}
