% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{filter_orthogroups}
\alias{filter_orthogroups}
\title{Filter the orthogroups of a tidygenomes object}
\usage{
filter_orthogroups(tog, ...)
}
\arguments{
\item{...}{Filtering expression to pass on to \code{\link[dplyr:filter]{dplyr::filter()}}}

\item{tg}{A tidygenomes object}
}
\value{
A tidygenomes object
}
\description{
This function applies \code{\link[dplyr:filter]{dplyr::filter()}} to the orthogroups table of a
tidygenomes object and subsequently filters other components of the
tidygenomes object to retain only the requested orthogroups.
}
