% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fix_pair_order}
\alias{fix_pair_order}
\title{Fix the pair order of a genome pair table}
\usage{
fix_pair_order(pairs)
}
\arguments{
\item{pairs}{A table with at least the variables genome_1 and genome_2}
}
\value{
A tibble
}
\description{
For each row in a genome pair table (table with at least the columns genome_1
and genome_2), this function assesses whether genome_1 < genome_2. If not, it
swaps the values for genome_1 and genome_2, as well as for other variables
whose names end in "_1" and "_2".
}
