% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{genes}
\alias{genes}
\title{Return a genes table}
\usage{
genes(tg)
}
\arguments{
\item{tg}{A tidygenomes object}
}
\value{
A genome table
}
\description{
This function returns the gene table of a tidygenomes object.
}
