% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{genomes}
\alias{genomes}
\title{Return a genome table}
\usage{
genomes(tg, extend = F)
}
\arguments{
\item{tg}{A tidygenomes object}

\item{extend}{Whether to add genome-related metadata}
}
\value{
A genome table
}
\description{
This function returns the genome table of a tidygenomes object, possibly
extended with extra genome-related data such as phylogroups.
}
