% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_tree}
\alias{get_tree}
\title{Return the tree of a tidygenomes object}
\usage{
get_tree(tg)
}
\arguments{
\item{tg}{A tidygenomes object}
}
\value{
A phylo object
}
\description{
Internally, the tree of a tidygenomes object is stored with n-numbers as tip
labels. This function returns a tree with genome names as tips.
}
