% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{ggtree_augmented}
\alias{ggtree_augmented}
\title{Call \code{ggtree()} and add metadata}
\usage{
ggtree_augmented(tg, ...)
}
\arguments{
\item{tg}{A tidygenomes object}

\item{...}{Extra arguments to pass on to \code{ggtree()}}
}
\value{
A ggplot object
}
\description{
This function calls the \code{ggtree} function of the ggtree package on the tree
component of a tidygenomes object and adds all available genome and node
metadata.
}
