% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{heatmap}
\alias{heatmap}
\title{Construct an heatmap to explore a pair variable}
\usage{
heatmap(tg, genome_label, distance, complete_pairs = T)
}
\arguments{
\item{tg}{A tidygenomes object}

\item{genome_label}{Variable (unquoted) from the genome table to use as
genome label}

\item{distance}{Variable (unquoted) from the pair table to plot on the
heatmap}
}
\value{
A ggplot object
}
\description{
This function constructs a heatmap of a given variable from the pair table.
Genomes are ordered along the phylogeny.
}
