% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importers.R
\name{inflate_pangenome}
\alias{inflate_pangenome}
\title{Inflate one of the pangenomes of a metapangenome}
\usage{
inflate_pangenome(tg_meta, tg_species, species)
}
\arguments{
\item{tg_meta}{A tidygenomes object containing the metapangenome}

\item{tg_species}{A tidgenomes object containing the species pangenome}

\item{species}{The genome in the metapangenome that represents the species to
inflate}
}
\value{
A tidygenomes object
}
\description{
This function inflates one of the pangenomes of what you could call a
"metapangenome": a pangenome where one or more individual genomes
represent(s) the entire pangenome of one species. The orthogroups of this
species are represented as individual genes in the metapangenome. This
function will replace the single species to be inflated by the individual
genomes of which is consists, and its "genes" with the complete species-level
orthogroups.
}
\details{
The genes in the metapangenome that belong to the genome (species) to be
inflated should correspond to orthogroups in the species pangenome.
}
