% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancers.R
\name{map_patterns}
\alias{map_patterns}
\title{Map presence/absence patterns to the tree}
\usage{
map_patterns(tg)
}
\arguments{
\item{tg}{A tidygenomes object}
}
\value{
A tidygenomes object
}
\description{
This function determines for each unique presence/absence pattern whether it
is a signature for a clade in the tree.
}
\details{
A variable \code{node} will be added to the pattern table.
}
