% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nodes_postorder}
\alias{nodes_postorder}
\title{Perform postorder tree traversal}
\usage{
nodes_postorder(
  tree,
  node = setdiff(tree$edge[, 1], tree$edge[, 2]),
  nodes = c()
)
}
\arguments{
\item{tree}{An object of the class phylo}
}
\value{
A numeric vector with node numbers
}
\description{
This functions returns the node numbers of a tree in order of postorder tree
traversal (tips first).
}
