% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nodes_preorder}
\alias{nodes_preorder}
\title{Perform preorder tree traversal}
\usage{
nodes_preorder(
  tree,
  node = setdiff(tree$edge[, 1], tree$edge[, 2]),
  nodes = c()
)
}
\arguments{
\item{tree}{An object of the class phylo}
}
\value{
A numeric vector with node numbers
}
\description{
This functions returns the node numbers of a tree in order of preorder tree
traversal (root first).
}
