% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{pairs}
\alias{pairs}
\title{Return a pair table}
\usage{
pairs(tg, extend = F, complete = F, add_reflections = F)
}
\arguments{
\item{tg}{A tidygenomes object}

\item{extend}{Whether to add genome-related metadata}

\item{complete}{Whether to add inverse pairs}

\item{add_reflections}{Whether to add reflexive pairs}
}
\value{
A pair table
}
\description{
This function returns the pair table of a tidygenomes object, possibly
completed with inverse pairs and possibly extended with extra
genome-related data such as phylogroups.
}
