% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pairs2matrix}
\alias{pairs2matrix}
\title{Convert pair table to matrix}
\usage{
pairs2matrix(pairs, object_1 = object_1, object_2 = object_2, measure, diag)
}
\arguments{
\item{pairs}{A table with pairwise comparisons}

\item{object_1}{Name of the variable specifying the first object of the pair}

\item{object_2}{Name of the variable specifying the second object of the pair}

\item{measure}{Name of the (dis)similarity variable}

\item{diag}{Something to fill the diagonal with}
}
\value{
A matrix
}
\description{
This functions converts a data frame where each row is a unique pairwise
comparison to a (dis)similarity matrix.
}
