% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{pangenome_matrix}
\alias{pangenome_matrix}
\title{Return a pangenome matrix}
\usage{
pangenome_matrix(tg)
}
\arguments{
\item{tg}{A tidygenomes object}
}
\value{
A matrix with the pangenome
}
\description{
This function returns the orthogroup content of the genomes in the form of a
matrix where the rows represent genomes, the columns represent orthogroups
and the cells contain the copy number or presence status of the orthogroups
in the genomes.
}
