% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{pansum}
\alias{pansum}
\title{Return amount of distinct genes, genomes and orthogroups}
\usage{
pansum(tg)
}
\arguments{
\item{tg}{A tidygenomes object}
}
\value{
Summary statistics of distinct genes, genomes and orthogroups
}
\description{
This function returns the total amount of distinct genes, genomes and orthogroups in
the tidygenomes object
}
