% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{plot_orthogroup_distribution}
\alias{plot_orthogroup_distribution}
\title{A density plot of the number of orthogroups per genome}
\usage{
plot_orthogroup_distribution(tg, limits = c(0.5, 0.75, 0.9, 0.95))
}
\arguments{
\item{tg}{A tidygenomes object}

\item{limits}{A vector of cumulative density limits to show
genome label}
}
\value{
A ggplot object
}
\description{
This function draws a density plot of the number of orthogroups per genome and indicates
a few predefined cumulative density percentages.
}
