% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_pangenome}
\alias{read_pangenome}
\title{Read pangenome from a file or folder}
\usage{
read_pangenome(path)
}
\arguments{
\item{path}{Path to a file or folder containing the pangenome}
}
\value{
A tibble with the variables gene, genome and orthogroup
}
\description{
This function reads a file or folder containing a pangenome, as created by
tools such as OrthoFinder, and returns a genes tibble with the genome and
orthogroup of each gene.
}
\details{
The \code{path} can currently only be an OrthoFinder Orthogroups folder (for
OrthoFinder >= v2.3.1), or Results folder (older versions). In the future,
this can be extended to output folders or files from other pangenome tools.
}
\examples{
\dontrun{
genes <- read_pangenome("Results_Sep06")
}

}
