% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_phylip_distmat}
\alias{read_phylip_distmat}
\title{Read a phylip formatted distance matrix}
\usage{
read_phylip_distmat(path, skip = 8, include_diagonal = T)
}
\arguments{
\item{path}{Path to a phylip formatted distance matrix file}

\item{skip}{Number of lines before the actual distance values start}

\item{include_diagonal}{Should the diagonal of the distance matrix be included?}
}
\value{
A tibble with the variables sequence_1, sequence_2 and distance
}
\description{
This function reads a phylip formatted distance matrix, as generated by e.g.
the EMBOSS tool distmat.
}
\details{
The number of lines to skip can be different for different phylip distance
matrix files.

For each row, the value of sequence_1 will be before the value of sequence_2
in sorting order.
}
