% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.R
\name{read_tidygenomes}
\alias{read_tidygenomes}
\title{Read tidygenomes object from a gzipped serialized JSON object.}
\usage{
read_tidygenomes(inf)
}
\arguments{
\item{inf}{Input file}
}
\description{
Read tidygenomes object from a gzipped serialized JSON object.
}
