% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importers.R
\name{remove_phylogroups}
\alias{remove_phylogroups}
\title{Remove phylogroups from a tidygenomes object}
\usage{
remove_phylogroups(tg)
}
\arguments{
\item{tg}{A tidygenomes object}
}
\value{
A tidygenomes object
}
\description{
This function undoes the effects of the function \code{\link[=add_phylogroups]{add_phylogroups()}}
}
