% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{root_tree.phylo}
\alias{root_tree.phylo}
\title{Root tree given three tips}
\usage{
root_tree.phylo(tree, tips, ...)
}
\arguments{
\item{tree}{An object of class phylo}

\item{tips}{Three tip labels}

\item{...}{
  Arguments passed on to \code{\link[ape:root]{ape::root.phylo}}
  \describe{
    \item{\code{phy}}{an object of class \code{"phylo"} or \code{"multiPhylo"}.}
    \item{\code{outgroup}}{a vector of mode numeric or character specifying the
    new outgroup.}
    \item{\code{node}}{alternatively, a node number where to root the tree.}
    \item{\code{resolve.root}}{a logical specifying whether to resolve the new
    root as a bifurcating node.}
    \item{\code{interactive}}{if \code{TRUE} the user is asked to select the node
    by clicking on the tree which must be plotted.}
    \item{\code{edgelabel}}{a logical value specifying whether to treat node
    labels as edge labels and thus eventually switching them so that
    they are associated with the correct edges when using
    \code{\link[ape]{drawSupportOnEdges}} (see Czech et al. 2016).}
  }}
}
\value{
An object of class phylo
}
\description{
This function roots a phylogenetic tree given three tip labels.
}
\details{
Tips a, b and c define exactly one internal node in the unrooted tree. The
tree will be rooted on the branch leading from this node to tip a.

By default, node/tip labels are interpreted as vertex labels, not edge
labels. When the labels actually represent edge labels (e.g., branch support
values), set edge_label to TRUE. This will make sure that the edge labels are
handled correctly, i.e. moved to the correct nodes. In that case, when the
tree is rooted on a non-tip branch, its edge label will be copied to both
child nodes of the new root node.
}
