% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{root_tree}
\alias{root_tree}
\title{Root tree given three genomes}
\usage{
root_tree(tg, root, genome_identifier = genome)
}
\arguments{
\item{tg}{An tidygenomes object}

\item{root}{Three tips that identify the root (see \link{root_tree.phylo})}

\item{genome_identifier}{Variable of the genome table that corresponds to the
given genomes}
}
\value{
A tidygenomes object
}
\description{
This applies \link{root_tree.phylo} to the tree component of a tidygenomes object.
}
