% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidygenomes-package.R
\docType{package}
\name{tidygenomes-package}
\alias{tidygenomes}
\alias{tidygenomes-package}
\title{tidygenomes: A Grammar of Genome Data Manipulation}
\description{
A toolkit for the organization and visualization of genome data, including pangenomes and phylogenetic trees.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SWittouck/tidygenomes}
}

}
\author{
\strong{Maintainer}: Stijn Wittouck \email{stijn.wittouck@uantwerpen.be}

}
\keyword{internal}
