% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{update_genomes}
\alias{update_genomes}
\title{Update the names of the genomes}
\usage{
update_genomes(tg, new_name)
}
\arguments{
\item{tg}{A tidygenomes object}

\item{new_name}{An expression that evaluates to unique names within the
genome table}
}
\value{
A tidygenomes object
}
\description{
This function updates the names of all genomes in all tables.
}
\details{
This function is deliberately not called "rename_genomes", because in the
tidyverse naming framework that would suggest that the function renames
variables in the genome table (which it doesn't do).
}
