% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{upset_plot}
\alias{upset_plot}
\title{Construct an "upset plot" to explore gene presence/absence}
\usage{
upset_plot(tg, genome_name, genome_col, genome_bold,
  color_scale = scale_color_brewer(palette = "Paired", guide = "none"),
  n = 50)
}
\arguments{
\item{tg}{A tidygenomes object}

\item{genome_name}{Name of a variable (unquoted) from the genome or
phylogroup table that should be used to name the genomes}

\item{genome_col}{Name of a variable (unquoted) from the genome or phylogroup
table that should be used to color the genome names}

\item{genome_bold}{Name of a logical variable (unquoted) from the genome or
phylogroup table that indicates which genome names should be in bold}

\item{color_scale}{Ggplot color scale generated by a scale_color_... type
function}

\item{n}{The number of presence/absence patterns to plot}
}
\value{
A ggplot object
}
\description{
This function constructs an "upset plot" showing patterns of gene familiy
presence/absence in order of pattern frequency (number of gene families
following the pattern). The genomes are ordered along their phylogeny.
}
