% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.R
\name{write_tidygenomes}
\alias{write_tidygenomes}
\title{Write tidygenomes object to a gzipped serialized JSON object.}
\usage{
write_tidygenomes(tg, outf = "tidygenome.gz")
}
\arguments{
\item{tg}{A tidygenomes object}

\item{outf}{Output file}
}
\description{
Write tidygenomes object to a gzipped serialized JSON object.
}
