% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancers.R
\name{add_gcd}
\alias{add_gcd}
\title{Add gene content distance}
\usage{
add_gcd(tg, method, binary)
}
\arguments{
\item{tg}{A tidygenomes object}

\item{method}{See \code{\link[vegan:vegdist]{vegan::vegdist()}}}

\item{binary}{See \code{\link[vegan:vegdist]{vegan::vegdist()}}}
}
\value{
A tidygenomes object
}
\description{
This function adds a variable gcd (gene content distance) to the pair table.
As the name suggests, these are distances between genomes based on shared
gene content. Various types of distances can be computed; the function uses
the vegan package for this.
}
