% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{filter_genomes}
\alias{filter_genomes}
\title{Filter the genomes of a tidygenomes object}
\usage{
filter_genomes(tg, ...)
}
\arguments{
\item{tg}{A tidygenomes object}

\item{...}{Filtering expression to pass on to \code{\link[dplyr:filter]{dplyr::filter()}}}
}
\value{
A tidygenomes object
}
\description{
This function applies \code{\link[dplyr:filter]{dplyr::filter()}} to the genomes table of a tidygenomes
object and subsequently filters all other components of the tidygenomes
object to retain only the requested genomes. It is also possible to filter on
variables of the phylogroups table, if present.
}
\examples{
genomes <- data.frame(genome = c("genome 1", "genome 2"))
tg <- as_tidygenomes(genomes)
tg <- filter_genomes(genomes, genome == "genome 1")

}
