% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{root_tree.phylo}
\alias{root_tree.phylo}
\title{Root tree given three tips}
\usage{
root_tree.phylo(tree, tips)
}
\arguments{
\item{tree}{An object of class phylo}

\item{tips}{Three tip labels}
}
\value{
An object of class phylo
}
\description{
This function roots a phylogenetic tree given three tip labels.
}
\details{
Tips a, b and c define exactly one internal node in the unrooted tree. The
tree will be rooted on the branch leading from this node to tip a.
}
