% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidygenomes.R
\docType{package}
\name{tidygenomes}
\alias{tidygenomes}
\alias{tidygenomes-package}
\title{tidygenomes: A Grammar of Genome Data Manipulation}
\description{
tidygenomes: A Grammar of Genome Data Manipulation
}
\details{
\code{tidygenomes} contains six types of functions: readers, importers,
handlers, enhancers, getters and plotters.
}
