% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{upset_plot}
\alias{upset_plot}
\title{Construct an "upset plot" to explore gene presence/absence}
\usage{
upset_plot(
  tg,
  genome_name = genome,
  genome_col = NULL,
  genome_bold = NULL,
  color_scale = scale_color_brewer(palette = "Paired", guide = "none"),
  n = 50,
  freq_min = NULL,
  barchart_height = 50,
  frequency_labels = F,
  phylogroups = F
)
}
\arguments{
\item{tg}{A tidygenomes object}

\item{genome_name}{Name of a variable (unquoted) from the genome or
phylogroup table that should be used to name the genomes}

\item{genome_col}{Name of a variable (unquoted) from the genome or phylogroup
table that should be used to color the genome names}

\item{genome_bold}{Name of a logical variable (unquoted) from the genome or
phylogroup table that indicates which genome names should be in bold}

\item{color_scale}{Ggplot color scale generated by a scale_color_... type
function}

\item{n}{The number of presence/absence patterns to plot}

\item{freq_min}{The minimum frequency of patterns to appear on the plot}

\item{barchart_height}{The relative height of the frequency barchart,
expressed as a number of rows from the main plot}

\item{frequency_labels}{Logical variable indicating whether the frequency
numbers should be plotted as labels on top of the barchart}

\item{phylogroups}{Should horizontal lines be plotted between phylogroups?}
}
\value{
A ggplot object
}
\description{
This function constructs an "upset plot" showing patterns of gene familiy
presence/absence in order of pattern frequency (number of gene families
following the pattern). The genomes are ordered along their phylogeny.
}
